feat: 🎸 parse code
