+++ src/common/parse/utils.ts
@@ -1,46 +1,9 @@
 import fs from 'fs-extra';
-import { endsWith, isEmpty, get, assign, keys, split } from 'lodash';
+import { isEmpty, get, assign, keys, split, filter, join, includes } from 'lodash';
 import path from 'path';
-import chalk from 'chalk';
 import { IProjectConfig, IActionHook, IInputs, IGlobalArgs } from './interface';
 import yaml from 'js-yaml';
-import { getYamlContent, getGlobalArgs } from '../../libs';
-
-async function validateTemplateFile(spath: string): Promise<boolean> {
-  if (isEmpty(spath)) return false;
-  try {
-    if (endsWith('json')) {
-      const data = fs.readJSONSync(spath);
-      return data.hasOwnProperty('edition');
-    }
-    if (endsWith(spath, 'yaml') || endsWith(spath, 'yml')) {
-      const data = await getYamlContent(spath);
-      if (isEmpty(data)) {
-        const filename = path.basename(spath);
-        throw new Error(
-          JSON.stringify({
-            message: `${filename} format is incorrect`,
-            tips: `Please check the configuration of ${filename}, Serverless Devs' Yaml specification document can refer to：${chalk.underline(
-              'https://github.com/Serverless-Devs/Serverless-Devs/blob/master/docs/zh/yaml.md',
-            )}`,
-          }),
-        );
-      }
-      return data.hasOwnProperty('edition');
-    }
-  } catch (error) {
-    return false;
-  }
-}
-
-export async function getTemplatePath(spath?: string) {
-  if (await validateTemplateFile(spath)) return spath;
-  const cwd = process.cwd();
-  const sYamlPath = path.join(cwd, 's.yaml');
-  if (await validateTemplateFile(sYamlPath)) return sYamlPath;
-  const sJsonPath = path.join(cwd, 's.json');
-  if (await validateTemplateFile(sJsonPath)) return sJsonPath;
-}
+import { getYamlContent } from '../../libs';
 
 export async function setupEnv(templateFile: string) {
   const spath = path.dirname(templateFile);
@@ -111,8 +74,8 @@ export function getActions(configs: IProjectConfig, { method, spath }): IActionH
   return hooks;
 }
 
-export function getInputs(configs: IProjectConfig, { method, args, spath }): IInputs {
-  const globalArgs = getGlobalArgs(args);
+export function getInputs(configs: IProjectConfig, { method, args, spath, serverName }): IInputs {
+  const argsObj = filter(args, (o) => !includes([serverName, method], o));
   const inputs = {
     props: configs.props,
     credentials: configs.credentials,
@@ -124,8 +87,8 @@ export function getInputs(configs: IProjectConfig, { method, args, spath }): IIn
       provider: configs.provider,
     },
     command: method,
-    args: globalArgs._args,
-    argsObj: globalArgs._argsObj,
+    args: join(argsObj, ' '),
+    argsObj,
     path: {
       configPath: spath,
     },
